Specified interfaces to files rather than one types file
* Moved interfaces to each individual file
- Removed types.ts
