+++ client/src/Components/ResourceTile.tsx
@@ -1,6 +1,26 @@
-import { HexGrid, Layout, Hexagon, Text, GridGenerator, HexUtils, Pattern } from 'react-hexgrid';
+import { HexGrid, Layout, Hexagon, Text, GridGenerator, HexUtils, Pattern, Hex } from 'react-hexgrid';
 import React from 'react';
-import { HexProp } from './types';
+import { Tile } from '@backend/types';
+
+/**
+ * An interface that provides strong typing to a resource tile's hexagon prop.
+ */
+export interface HexProp {
+    /**
+     * A hexagon object representing this particular tile's placement on the grid.
+     */
+    hex: Hex;
+
+    /**
+     * The numbered index of this tile -- as in, its index placement on the grid.
+     */
+    index: number;
+
+    /**
+     * The backend information related to this hexagonal tile.
+     */
+    tile: Tile;
+}
 
 /**
  * An individual component used by the gameboard as a tile. Each resource tile has a
